Formatting the README and condensing
Hello. 

I just formatted the headings and line numbers used in the instructions. Then I took out the duplicated set of instructions, made the commands more consistent to follow and indicated what steps are by choice.

I'm assuming in the *development* section that the github master branch will be kept stable otherwise the install instructions should change to reflect that. 

Also I'm not on a linux box at the moment, someone should check I can pipe wget to tar like that and create the trunk folder. I don't use svn so I check the doc to reverse the `co` alias into the command it reflect for clarity.