Fixed ubuntu client install
Reduce the number of tries and length of time to attempt to wget the FOG site.

The wget command has a default -T value of 900 seconds and -t has a default of 20 tries. Doing this 20 times at 900 seconds each will take a long time. That could mean someone having issues could be sitting for 360,000 seconds or 4.1 days.

This fix reduces maximum wait time to 2.5 minutes.