All 4 FOG services now have a persistent wrapper.
This deals with the race condition of the service starting
before the database server is available.  Since this rears
its head in the require_once() of basic FOG setup, the
looping in the services themselves are not able to recover.

This was most notably a problem in Ubuntu 14.04.  The solution
was tested there.  It relies on PHP having pcntl_* and posix_*
available.

Installers for ubuntu/redhat/arch have been modified to install
configuration for this addition.