Ensure hostname changer removes the semicolon from the OU in the case OU is setup as a select box.
Update multicast tasks to layout the command so that the first file in
the list gets the defined maxwait time while the other files in the list
wait for only 10 seconds.

Add the session write close function to the getfiles as getfiles is an
ajax call.  We don't want log files to cause blocking on changing to
other files.