Make checkout it's own function in the task queue class and use it to checkout tasks that have completed.  Only set for deploy and non imaging task completions for now.
Setup the multicast task with the list in a more suitable fashion to
ensure variables get populated properly.

FOG Settings used to make 1 save for each item everytime you pressed
save.  This could take a little while.  Move to using the insert_batch
method to perform the update/save to ensure speed is commonality are
used.