Fix report date logging
Fix report logging that deals with dates to use the proper date constructs.

Ultimately it allows items that end days further ahead which makes sense.  We're searching for stuff that had a start date for the dates we want to see, so why should we not display tasks that started on one date, but ended on another?