Push RC 32.
This fixes snapin and image primary group setting. Particularly, it will define the primary group properly in the case one is not already defined. It will not require double entry to set the primary group after fresh import either.

Should address issue with slow FTP Image size getting at least partially. It was slowed, quite significantly, as it was translating the returned data into a double. This conversion took a bit longer as there was numerical and alphabetical data and we needed to parse all of it.