File permission fixes.
File permissions on images and snapins should be set to 777 as is the case under normal conditions. The Installer should set the images folder, for sure, on each install rather than only if some other change had to occur.

When snapins are uploaded, make sure they have the same permissions level.  When an image is captured set that image to 777 as well.