Fix url callers so IE works.
Firefox, Chrome, Safari, Edge, and Opera worked fine.  IE NAH, not so much.
Basically the about page wouldn't display the version information properly to the user
because the url it would self call would simply be appended to the current URL.
The other browsers worked properly because they understood the relative nature, but no
NOT IE.