Fix replication issue.
Replication for both images and snapins was broken during the switch to methodized calls for looping.
In regards to snapins and images, we have to iterate the snapins and images iteratively twice.
Forgetting this, the issue occurred because the secondary variable did not exist.  Reset to use
the variable method (which limits sql calls.)