Fix installer checks.
The installer doesn't test the boot type used for system services. While
systems might have the systemctl command, it does not mean systemd was
what iniated the services.  Add a test for the boot type (init/systemd)
and use it as a secondary means to ensure we use the correct service handler
for the system on boot. Also, we should allow the user to define if they
want to use systemctl or not.  So this fix adds the init/systemd boot model
tester and checks if the systemctl is set prior or not.