Add module shortnames for testing.
The new client will report non-enabled modules but there may be cases
where the non-enabled modules still get run by the client.  We should've had
a checker to prevent "actions" on this but it didn't exist initially.

Correct this behaviour by making the client elements check themselves. Even
if the client is requesting actions to perform when it probably shouldn't be
we will make the PHP side handle this and prevent the action at an individual
level.