DB Errors were not printing for schema updates.
Schema updates would always return as true because the object was
being returned instead of a boolean as being checked. This would make
it ever more difficult to track down bugs in the sql statements
especially considering there are over 3k lines.
In older versions, even those that did display errors, in some cases
could still blast through and update the schema version. This left
potential unset database items that could be referenced elsewhere
in the system. Due to this the schema version will no longer be updated
past and should stop trying to execute beyond that point.

Address this by:
1. Store the error message into the current property.
2. Run through all the schema updates and collect any errors. Break if error
  is found.
3. Return an http response code that will make updating the database
  from the installer fail to execute.
4. The schema version will not be updated past that point so a bug can be
  reported and hopefully fixed relatively quickly.

Among this schema rethought, cleanup the myisam conversion as it really
shouldn't be needed. It was for some time, but we really shouldn't test
this every time you click a link (it creates more IO to read the tables,
 check types, and iterate over the types to change them to myisam).

Cleanup the report image and snapin log date presentation. Instead of calling
a large sql to group an array to filter, just get the min start date and display
all dates in between (even if nothing happened on those dates).