Rework Dashboard page.
The dashboard page is most often the least kept on page
considering most people go in to perform actual tasks. One
of the biggest prolems with this page, however, is it was
highly memory intensive. This memory is (yes it is still present)
caused by the 30 day graph. We have to get all thirty days and gather
the counts individually. This caused a 2 Mb initial display use 7 times
more memory because we had to wait for the page to iterate each date
and make 30 separate calls.

In an attempt to improve the load in performance the page is now immediate
when compared earlier. All the dynamic parts (bandwidth, disk usage, and client
count) are still dynamic, but this also moves the 30 day gathering into a
kind of dynmaic state. Instead of causing a block on the page to gather
the 30 days of info, we use an ajax call to get the information.