Fix user issues.
For the most part, the user login is working but found a few oddities particularly if working with the ldap plugin. This addesses all those issues and in particular allows
the ldap plugin to work with mobile users as well.

TODO: Have the delete elements work if the users haven't logged in or don't exist
first.