Use sgdisk to verify partition.
While we had the piped in yes statement, this did not prevent issues.
The yes statement was not doing anything with the drive and did allow
imaging to capture and deploy like all was well.  While the data is indeed
on the disks, if a disk is in an unexpected layout state it would remain
that way causing booting to fail potentially.

Attempt to fix this (still needs good testing to verify) by verifying the
partition layout through the use of `sgdisk -v <disk>` and if there is
and error, have the system try to perform fixparts on the disk.