Fix storage node potential issue.
Users might want to use / as their image path or ftp path.

Using / would cause the rtrim to active leaving the ftppath or imagepath fields blank.  Because these fields are required in the storage node list it, the nodes with either and/or of these fields with / would present as invalid nodes. This pushes that up to correct that issue.