More bug fixes.
1. Host MAC table should not allow duplicate MACs ever.
This is addressing the hostMAC table to only allow a single mac address rather than associating duplicate mac's with multiple different hosts. While this is kind of allowable, really we only want one mac associated.

2. Service page editing should return to item being edited.
This is addressing updates from the service page redirecting to the home page on any update. It also address the form actions to pass the subs to the edit method so the updates can actually take effect. Base updates where enabling default and main item worked but if there was additional entries the form missed the edit item.

3. Initial submission to address legacy client functionality.
The 1.3.1/1.3.2 series to address the new client failed to recognize the legacy client. This corrects the action to ensure legacy functionality remains. Still testing to ensure all legacy components are at least functioning from a logging standpoint.