BCD Containing partition should be fixed.
BCD Containing partitions should be setup as 'fixed' in state. This
is more a test for a theory, but the code used to handle bcd partitions
and anything below would become a fixed size. What happens if the bcd
is on a two partition setup (recovery and main data?) This would make
the whole disk not resizable.