Schema version redirect.
The redirect should only happen if the reporting SCHEMA is greater than the current schema. If the current schema is less than the reporting schema we know an update will be needed and the page will redirect. The error, here, is it was redirecting if the reporting schema was not equal to the current schema. If you downgraded the version that had a smaller reporting schema than the current schema it would put your system into an infinite redirect loop which ultimately breaks
access to the GUI. Address by checking if reporting schema is larger than the current schema only.