Add multicast rendezvous address.
This setting is typically uncommon, but could be needed in some environments.
FOG's implementation only used the rendezvous from the receiver, not the sender,
but it's usage was not correct. As this is a rendezvous point, both sides of the
multicast traffic should look at this point. This was not occurring as only
the client was looking at a rendezvous of the servers ip address. This, itself
was not working properly.