Add quick register feature.
This just adds a more implicit means to tell if a host should or should not
be imaged when registration is complete. Relying on the image id solely as
the tester means you could auto associate an image with a host and just return.
It was forcing a system to image on completion of the registration task. This
just makes it so it can be more granularly controlled.