Tweaks to init and replication checking.
There was an issue with out the partition type was being applied
to the disk during imaging. While unsure at this point, but it could
cause issues to the rest of the booting process in special cases as the
types wouldn't match what the originating OS is expecting it to be.

Use the first and last 10 megabytes of files to gather the hash instead
of just the last. This should prevent issues and more accurately detect
changes to the files themselves.