Group should allow multiple snapin tasks.
Ensure groups page can task snapins at an individual level multiple times. If one host in the grouping is currently in an imaging task, it will not task any other hosts. To do this, add elements to get "tasking" identifiers from the task type page to filter out the task count more easily. Due to the potential of a snapin job having multiple jobs, use the min snapin job to present first (so if you wanted to task a snapin job in a specific order.)