Add security elements.
Add API Enabler from GUI.
Add API Token and reset capabilites to GUI.
Check the passed token and the set token to ensure they match.
Check the user/pass information from Basic authentication type.
If API is not enabled, redirect user to home page.
If Token is not set or not correct, show forbidden.
If User/Pass is not set or correct, show unauthorized.