On scheduled tasks add hover text.
The scheduled task page woul display the task name, but that name isn't always useful.
Viewing full information on this will include a hover text. This text will now display
a snapin's name to be installed in the case of Single Snapin task. This will show up as a
tooptip when hovering within the tasks name.

Fix display of snapin tasking on main active page when it is tasked.

Fix task creation to not just set the snapin type to All snapins. Perform better logic to test if the deploySnapin is already tasked or not.