Work plugin system to fog2 gui.
Initial work toward getting plugins to operate under the new GUI.
Currently just working out the location plugin and more specifically for
group page.

TODO: Get javascript/jquery to be pluginable for ajax calls.

TODO: Rethink how to approach passing the current object as the
tabFields function is statically called, we can't simply check the
'$this->childClass' or '$this->obj' elements. For now, just added
mechanism to method of passing the object, with a default value preset.
Default is -1 which tells the method to load the current node + id as
the current object to pass in.