Site Plugin Enhancements
1. We needed a way to dynamically add plugin javascript files. Adjusted
now to use a separate hook rather than try to load it all at once. This
way we can inject any javascript how we see fit. Plugins are now
capable of doing this.

2. Most of the changes are in regards to the site plugin after above.
This is not by any means the only work needed, but the majority of this
work is dealing with that as you can see here.