Power Management and Location plugin
1. Work on Powermanagement tabs to become function. Groups should be
done as we do not need to list individual power management items. Hosts
still need the power management list populated/presented. The adding
element is functional though.

2. Get location plugin mostly configured with the jquery code, still
needs work but mostly coded out for now.

3. Plugins should no longer need to check if they're loaded inside each
method when working between the hooks. This is because the methods and
routes to those methods won't be available if the plugin is no longer
present or just not installed.  This can be handled in each hooks
__construct method. Right now testing is limited to only Site and
Location Plugins.