h3 to h4
1. Move all h3 to h4 tags for consistency.

2. Move mac addresses to their own tab on host page. Still need to get
the jquery operation for it, but adding new macs is already working. (It
also updates the table when a new address is added successfully.) Of
note, a mac address is still required during add, just when editing a
host will you see the macs listing.

3. Add mac address mask to filter characters appropriately.

4. The default code for ping status on hosts is showing a function
rather than the default value, this should correct that action.