Schema edits mostly.
1. Get Schema updater to be new gui friendly, including the relevant
jQuery. This means errors are now stored to the FS under the basepath of
the fog folder. The filename containing the sql errors and relevant
messagages is fog_schema_update_error.log.

2. Add makeLabel, makeInput, and makeFormTag methods in fogpage. These
are static methods for ease. It allows us to abstract the generation of
field labels and field inputs so things are commonized across the board.
(All inputs are generated in the same fashion/order rather than leaving
potential differences and errors up to the user.

3. With the new methods from above, the only page strictly using these
new methods is the user management page. This will be changed out for
all field elements with a bit more time of course. Users page now uses,
makeButton method for all buttons being used. It uses makeFormTag for
all form areas. It uses makeLabel for all field label areas. It also
uses makeInput for all input areas (Our presentation is <label> <input>
in very simple terms.)