User Updates Mostly.
1. Update FOG Page makeTabUpdateURL method only set an id to the url if
an ID is passed, it should not be a requirement for the url as we can
test tabs even on items without an pre-existing object.

2. Make user page contain it's own export method. Include the export
list for datatables. We will present information in a selected fashion
to the user. Rather than using fog's own ReportMaker, as it has many
problems. While we can use it for sql dumps, it probably should be
retired when working with PDF/CSV, especially considering Datatables has
these items and more readily available.

3. Fix fog.common.js datatable opts element so it can define things
properly when we're calling registerTable from other js files.