Dashboard tweaks.
1. Remove the old bandwidth defining item. While it could be
reimplemented, the primary reason for it was due to performance issues.
Nothing seemed to have fixed this until this most recent refactoring.
Because of this, our time should be set back to 1 second intervals, for
proper "realtime" data.

2. Update the on|off realtime element so that we don't see updates, but
data is still collected. When the on is pressed again, or another option
chosen (Transmit to recieve, 2 minutes to 5, 10, etc...) it should start
displaying again as well as show the missed data when it was paused.