Fixup underscore issue with JS.
1. fog.common.js now uses _.defaults which was causing issues. First
issue appears to have been with the database.min.css file that was being
used. It was around 2.5 MB so that leads me to think many files were
concatenated within it. This, itself, isn't a big problem, but the order
apparently matters. Rebuilt the datatables.min.js file using datatables
download builder. This corrected this problem. However, it did not have
the underscore js library built in for some reason.

2. Downloaded the underscore js library and applied it, which fixed the
issues with _.defaults entirely (it appears).

3. Put pending macs on the host sub menu element since mac addresses are
always associated to some sort of host device. Javascript has been
created for handling it as well (seeing as it is very similar to pending
hosts, but completely untested so far.)