Many fixes and changes.
Seems large, but mostly just tidying up.

Moved FOGFTP setting to __get/__set so we can call on variables for the
class directly as opposed to forcing a function call.
Fix the isavailable checks to use stream_socket_client with a timeout of
0. This seems to be providing a much faster loading page for things that
used to take eons if multiple hosts were unavailable.