Add alerts more nicely.
For the dashboard page, we need to account for the error part of ajax to
set up an alert as well. If a page can't be reached we would never see
an error, but the url called would return a 400. So use the error
processing to display alerts a bit more directly.