Finish up work with ipxe page elements.d10d
Fix issue with FormData/Serialize as they are needed at different
points. In the case of a form that's multipart/form-data we need to use
FormData. With that, the contentType should be false, and processData
should be false. If the encoding type is
application/x-www-form-urlencoded, we do need to processData, and we
need to set the content type to the same. I've added this to the apiCall
element. if processData is true, content type will be set to
application/x-www-form-urlencoded, otherwise it will be set to false.
processData will simply be set to whatever the passed variable is.

As ipxe page is in place, remove the trailing bits from the
configuration page. The main settings of the menu are still under
configuration, though I may move them into a sub of the ipxe menu system
later on.