Install fixes, Refind adjust
1. Adjust Refind.conf to set netboot,internal,hdbios as in UEFI having
internal causes the machine to reset nvram to use the first hdd rather
than maintain PXE booting capability.

2. Modify input.sh so that if an interface is implicitly defined, the
IPAddress and Subnet mask get set still. Before it only set these fields
if interface wasn't set. Do so by wrapping the interface checking into a
function. Outside of the function, test the validity of the IP if its
not already defined. After the IP is tested, try to get the subnet mask.

3. Modify the schema file to set the fresh install of a storage node's
image location to that set to storageLocation variable. (It will set
this to the storageLocation every install too, but it's only valid
during fresh install and creating the DefaultMember storage node.)